bugfix to prevent creation of HTML files that the build has not already created

+++ jenkins/inject_date.sh
@@ -6,16 +6,8 @@
 # Do not use the -time switch on public documents.
 # If you want to get the date from the timestamp instead of the commit, use -file
 
-echo ===start inject_date.sh===
-for i in `find  -name "*.dita" -not -path "./publiccloud/api/*"`
-do
 
-    j=`echo $i | sed 's|\.dita$|\.html|'`
-    fullpath=`echo $j | sed 's|\.\/|./out/webhelp/|'`
-    echo $fullpath
-    #DATE=`git log -1 --date=short --pretty=format:%ad $i`
-
-    if [ "$1" == "-file" ]
+   if [ "$1" == "-file" ]
     then
     
     DATE=`stat --format=%y $i`
@@ -30,15 +22,28 @@ do
     else
     	PRETTYDATE=`date -d"$DATE" +'%d %b %Y'`
     fi
-
-    sed -i "s|<\/h1>|</h1><p class=\"heliondate\">Last updated: $PRETTYDATE<a href=\"\" class="xref" style=\"float:right\" onclick=\"window.print()\">Print this page</a> </p>|" $fullpath    
 	
-	echo starting from 	
-	pwd
-	echo before time change
-	stat --format=%y   $fullpath 
-	touch -d "$PRETTYDATE"  $fullpath
-echo after time change	
-     stat --format=%y   $fullpath 
+	
+
+echo ===start inject_date.sh===
+for i in `find  -name "*.dita" -not -path "./publiccloud/api/*"`
+do
+
+    j=`echo $i | sed 's|\.dita$|\.html|'`
+    fullpath=`echo $j | sed 's|\.\/|./out/webhelp/|'`
+    echo $fullpath
+    #DATE=`git log -1 --date=short --pretty=format:%ad $i`
+
+    if [ -e $fullpath ]
+	then
+		sed -i "s|<\/h1>|</h1><p class=\"heliondate\">Last updated: $PRETTYDATE<a href=\"\" class="xref" style=\"float:right\" onclick=\"window.print()\">Print this page</a> </p>|" $fullpath    
+		echo starting from 	
+		pwd
+		echo before time change
+		stat --format=%y   $fullpath 
+		touch -d "$PRETTYDATE"  $fullpath
+		echo after time change	
+		stat --format=%y   $fullpath 
+	fi
 done
 echo ===end inject_date.sh===
